+++ modules/hanna-react/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- fix: Pass `id` and other HTML props to static (span) `TagPill`s
 
 ## 0.10.62